requirements.txt
# Lease Management System - Python Dependencies

# Core Web Framework
Flask>=2.0.0
flask-cors>=3.0.0

# Authentication & Security
bcrypt>=4.0.0
cryptography>=3.4.0

# Date & Time Utilities
python-dateutil>=2.8.0

# File Handling
Werkzeug>=2.0.0

# HTTP Requests (for tests)
requests>=2.25.0

# PDF Processing (for AI extraction features)
pdfminer.six>=20201018
pdfplumber>=0.7.0
pypdf>=3.0.0

# AI extraction with Google Gemini
google-generativeai>=0.3.0

# Optional: For OCR (scanned PDFs)
# pdf2image>=1.16.0
# pytesseract>=0.3.10

# Database (SQLite is built-in, but can add ORM if needed)
# sqlalchemy>=1.4.0  # Optional: if migrating to ORM

